Change float to double + Display heuristic value of best node

+++ core/src/com/abalone/game/gameTree/BoardState.java
@@ -26,7 +26,7 @@ public class BoardState {
         return turnsFinder.getTurns();
     }
 
-    public float getValue(){
+    public double getValue(){
         return heuristics.value;
     }
